backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aioredis==2.0.1
websockets==12.0
openai==1.3.0
deepgram-sdk==3.0.1
twilio==8.10.0
requests==2.31.0
httpx==0.25.0
pinecone-client==3.0.0
langchain==0.1.0
langchain-openai==0.0.2
python-multipart==0.0.6
stripe==7.4.0
